Revert "Revert "feat: rename geometry module to obstacle""
This reverts commit 234a4e10abc930bee4358ff8b7daf5bf275da8b6.
