+++ src/euler_fluid.rs
@@ -1,6 +1,6 @@
 pub mod definition;
 pub mod fluid_bind_group;
-pub mod geometry;
+pub mod obstacle;
 pub mod render_node;
 pub mod setup_components;
 
@@ -24,7 +24,7 @@ use definition::{
     PressureTextures, SimulationUniform, VelocityTextures,
 };
 use fluid_bind_group::FluidPipelines;
-use geometry::Velocity;
+use obstacle::Velocity;
 
 use render_node::{EulerFluidNode, FluidLabel};
 
@@ -186,7 +186,7 @@ impl Plugin for FluidPlugin {
 }
 
 fn update_geometry(
-    query: Query<(&geometry::Circle, &Transform, &Velocity)>,
+    query: Query<(&obstacle::Circle, &Transform, &Velocity)>,
     obstacles: Res<Obstacles>,
     mut buffers: ResMut<Assets<ShaderStorageBuffer>>,
 ) {
