+++ src/euler_fluid/obstacle.rs
